+++ services/GateWay/include/global/Global.hpp
@@ -74,8 +74,17 @@ constexpr std::size_t REDIS_DB_INDEX = 0;        // Redis 数据库索引
 constexpr std::size_t REDIS_MAX_POOL_SIZE = 16;  // Redis 最大连接池大小
 constexpr std::size_t REDIS_TIMEOUT = 3;         // Redis 连接超时时间
 
+constexpr auto JWT_DEFAULT_SECRET = "ChatRoom-Secret-Key-2025";
+constexpr auto JWT_ISSUER = "ChatRoom-GateWay";
+constexpr auto JWT_EXPIRATION_TIME = 7 * 24;  // JWT 过期时间 7天
+
+constexpr std::size_t RATE_LIMIT_WINDOW_SIZE = 60;    // 窗口大小 60秒
+constexpr std::size_t RATE_LIMIT_MAX_REQUESTS = 100;  // 每分钟最大请求数
+
 constexpr const char* VERIFY_CODE_PREFIX = "verify_code_";  // 验证码在 Redis 中的键前缀
 constexpr const char* USER_INFO_PREFIX = "user_info:";      // 用户信息前缀
+constexpr const char* RATE_LIMIT_PREFIX = "rate_limit:";    // 限流前缀
+
 constexpr const char* DEFAULT_AVATAR_URL =
     "http://14.103.206.66:8888/group1/M00/00/00/oYYBAGk3_NKAO9_qAAFp-FKCqaY435.jpg";  // 默认头像 URL