+++ services/GateWay/src/core/domain/vo/user/user_vo.hpp
@@ -26,6 +26,7 @@ struct CORE_EXPORT UserLoginVO
   std::string host;
   std::int16_t port;
   std::string token;
+  std::string jwt_token;
 
   [[nodiscard]] Json::Value ToJson() const
   {
@@ -34,6 +35,7 @@ struct CORE_EXPORT UserLoginVO
     root["host"] = host;
     root["port"] = port;
     root["token"] = token;
+    root["jwt_token"] = jwt_token;
     return root;
   }
 };